requirements.txt
fastapi[standard]
jinja2>=3.1.3
mysql-connector-python>=8.3.0  # Pure Python connector
python-dotenv>=1.0.0
cryptography>=42.0.5        # SSL support
numpy
bcrypt
